Restrict http-client auth to a single origin
This fixes a security issue that meant any config-URL supplied to
mast would have the artifactory credentials attached in requests
sent to it, potentially revealing them to a third-party.
It introduces a new, mandatory, config option, `ARTIFACTORY_ORIGIN`,
which is compared against the origin (scheme+netloc, using urlparse
terminology), and only when the request is to the same origin will
the credentials be attached.