Update to python 3.9, use ubuntu2004 CI image
The ubuntu1804 CI image is deprecated[1], and the oldest python version
available in the newer CI image is 3.9. Upgrade to Python 3.9, and sort
out several transitive dependency incompatibilities to make that work.

Several dependencies are pinned at older versions and should be
upgraded, but I think it is better to do that separately.

[1]: https://semaphoreci.com/blog/deprecating-ubuntu-18.04-in-semaphore